Created landlord.php with addApartment() HALF working. Read further...
landlord.php is newly added to hold addApartment() method which is called from the to process the 'Add Apartment' form from the UI.

ApartmentDB.php - addApartment($aprt) has been modified for better readability and refit to work the new columns + type checking.

TEST.php - not everything is implemented so this file exist purely to test other features that require something else that's still in progress.